+++ frontend/src/App.jsx
@@ -6,6 +6,7 @@ import SearchList from "./routes/SearchList";
 import CoupleCodeMatch from "./routes/ConnectCouple";
 import Place from "./routes/Place";
 import MyPage from "./routes/MyPage";
+import RegisterUser from "./routes/RegisterUser";
 
 function App() {
     return (
@@ -22,6 +23,7 @@ function App() {
 
                 
                 <Route path="/mypage" element={<MyPage />} />
+                <Route path="/registerUser" element={<RegisterUser />} />
                 <Route path="/connect-couple" element={<CoupleCodeMatch />} />
             </Routes>
         </Router>
